fix all known bugs/issues in github

+++ react-app/src/components/NotebooksDropdownMenu/index.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import { useState, useRef, useEffect } from 'react';
+import React, { useEffect } from 'react';
+import { useState, useRef } from 'react';
 import {useDispatch, useSelector} from 'react-redux';
 import {NavLink} from 'react-router-dom';
 import {BsPlusCircle, BsDashCircle, BsTrash, BsPencil, BsX} from 'react-icons/bs'
@@ -17,29 +17,56 @@ const NotebooksDropdownMenu = () => {
     const [showNotebookForm, setShowNotebookForm] = useState(false);
     const [showEditNotebookForm, setShowEditNotebookForm] = useState(false);
     const [showNoteForm, setShowNoteForm] = useState(false);
+    const [currentEle, setCurrentEle] = useState(null);
     const dispatch = useDispatch();
 
     const user = useSelector(state => state.session.user)
     const notebooks = useSelector(state => state.notebooks.notebooks);
 
-    const handleClick = () => setIsOpen(!isOpen);
+    const handleClick = () => setIsOpen(true);
+
+    const handleEditNotebookClick = (id) => {
+        setShowEditNotebookForm(true);
+        setCurrentEle(id)
+    }
 
     const deleteNotebook = (notebookId) => {
         let userId;
         if (user) {
             userId = user.id
         }
+        let answer = window.confirm(`Are you sure you want to delete this notebook?`)
+        if (answer) {
         dispatch(notebookActions.deleteNotebook(userId, notebookId))
+        } else {
+            return;
+        }
     }
+
+    useEffect(() => {
+        function handleClickOutside (e) {
+            if (!dropdownRef.current) {
+                return;
+            }
+
+            if (!dropdownRef.current.contains(e.target)) {
+                setIsOpen(false);
+            }
+        }
+        window.addEventListener('click', handleClickOutside);
+        return () => {
+            window.removeEventListener('click', handleClickOutside);
+        }
+    }, [dropdownRef, isOpen])
 
 
     return (
-        <div className={styles.notebook_menu__container}>
+        <div ref={dropdownRef} className={styles.notebook_menu__container}>
             <button onClick={handleClick} className={styles.notebook_menu__trigger}>
                 <span>Notebooks</span>
             </button>
             {isOpen && 
-            <div ref={dropdownRef} className={`${styles.notebook_menu} ${isOpen ? styles.active : styles.inactive}`}>
+            <div className={`${styles.notebook_menu} ${isOpen ? styles.active : styles.inactive}`}>
                 <button className={styles.add_notebook} onClick={() => setShowNotebookForm(!showNotebookForm)}>Create Notebook {!showNotebookForm ? <BsPlusCircle /> : <BsDashCircle />}</button>
                 {showNotebookForm && 
                 <div className={styles.form__wrapper}>
@@ -49,20 +76,20 @@ const NotebooksDropdownMenu = () => {
                 }
                 {notebooks?.map((notebook) => (
                     <div  key={notebook.id}>
-                    <div className={styles.notebook_container}>
-                        {!showEditNotebookForm && 
-                        <>
-                        <ul className={styles.notebook_title}>{notebook.title}</ul>
-                        <button className={styles.edit_notebook} onClick={() => setShowEditNotebookForm(true)}><BsPencil /></button>
-                        <button className={styles.delete_notebook} onClick={() => deleteNotebook(notebook.id)}><BsTrash /></button>
-                        </>
-                        }
-                        {showEditNotebookForm && 
-                        <div className={styles.form__wrapper}>
-                        <EditNotebookForm setShowEditNotebookForm={setShowEditNotebookForm} currentTitle={notebook.title} notebookId={notebook.id}/>
-                        <button onClick={() => setShowEditNotebookForm(false)} className={styles.close}><BsX /></button>
-                        </div>}
-                    </div>
+                        <div className={styles.notebook_container}>
+                            {!showEditNotebookForm && 
+                            <>
+                            <ul id={notebook.id}className={styles.notebook_title}>{notebook.title}</ul>
+                            <button id={notebook.id} className={styles.edit_notebook} onClick={(e) => handleEditNotebookClick(e.target.id)}><BsPencil /></button>
+                            <button className={styles.delete_notebook} onClick={() => deleteNotebook(notebook.id)}><BsTrash /></button>
+                            </>
+                            }
+                            {showEditNotebookForm && notebook.id === currentEle &&
+                            <div className={styles.form__wrapper}>
+                            <EditNotebookForm setShowEditNotebookForm={setShowEditNotebookForm} currentTitle={notebook.title} notebookId={notebook.id}/>
+                            <button onClick={() => setShowEditNotebookForm(false)} className={styles.close}><BsX /></button>
+                            </div>}
+                        </div>
                         <button className={styles.add_note} onClick={() => setShowNoteForm(!showNoteForm)}>Create Note {!showNoteForm ? <BsPlusCircle /> : <BsDashCircle />}</button>
                         {showNoteForm && 
                         <div className={styles.form__wrapper}>
@@ -72,9 +99,9 @@ const NotebooksDropdownMenu = () => {
                         
                         <div className={styles.notes_container}>
                         {
-                            Object.values(notebook.notes).map(note => {
+                            Object.values(notebook.notes).reverse().map(note => {
                                 return (
-                                    <NavLink className={styles.note_nav}to={`/notebooks/${notebook.id}/notes/${note.id}`} key={note.id}>{note.title}</NavLink>
+                                    <NavLink onClick={() => setIsOpen(false)} className={styles.note_nav}to={`/notebooks/${notebook.id}/notes/${note.id}`} key={note.id}>{note.title}</NavLink>
                                 )
                             })
                         }
